Выносит auth и getLayoutState в файлы const и utils/common соответственно
