+++ src/components/layout/layout.tsx
@@ -3,28 +3,7 @@ import { Outlet, useLocation } from 'react-router-dom';
 import { AppRoute, AuthorizationStatus } from '../../const';
 import Header from '../header';
 import Footer from '../../pages/favorites/components/footer';
-
-const getLayoutState = (pathname: AppRoute) => {
-  let rootClassName = '';
-  let linkClassName = '';
-  let title = '';
-
-  if (pathname === AppRoute.Root) {
-    rootClassName = ' page--gray page--main';
-    linkClassName = ' header__logo-link--active';
-  } else if (pathname === AppRoute.Login) {
-    rootClassName = ' page--gray page--login';
-    title = ': authorization';
-  } else if (pathname === AppRoute.Favorites) {
-    title = ': favorites';
-  } else if (pathname === AppRoute.Offer) {
-    title = ': offer';
-  } else {
-    title = ': page not found';
-  }
-
-  return {rootClassName, linkClassName, title};
-};
+import { getLayoutState } from '../../utils/common';
 
 type LayoutProps = {
   authorizationStatus: AuthorizationStatus;
