+++ src/utils/common.ts
@@ -1,3 +1,27 @@
+import { AppRoute } from '../const';
+
+const getLayoutState = (pathname: AppRoute) => {
+  let rootClassName = '';
+  let linkClassName = '';
+  let title = '';
+
+  if (pathname === AppRoute.Root) {
+    rootClassName = ' page--gray page--main';
+    linkClassName = ' header__logo-link--active';
+  } else if (pathname === AppRoute.Login) {
+    rootClassName = ' page--gray page--login';
+    title = ': authorization';
+  } else if (pathname === AppRoute.Favorites) {
+    title = ': favorites';
+  } else if (pathname === AppRoute.Offer) {
+    title = ': offer';
+  } else {
+    title = ': page not found';
+  }
+
+  return {rootClassName, linkClassName, title};
+};
+
 //получение случайного элемента массива
 function getRandomArrayElement<T>(items: readonly T[]): T {
   return items[Math.floor(Math.random() * items.length)];
@@ -31,4 +55,4 @@ function ucFirst(str: string):string {
   return str[0].toUpperCase() + str.slice(1);
 }
 
-export { getRandomArrayElement, createIdGenerator, getRandomInteger, ucFirst };
+export { getRandomArrayElement, createIdGenerator, getRandomInteger, ucFirst, getLayoutState };
